Handle exceptions for the delete and update endpoints

+++ src/main/java/com/highlander/DesafioCrudDeClientes/services/ClientService.java
@@ -3,11 +3,15 @@ package com.highlander.DesafioCrudDeClientes.services;
 import com.highlander.DesafioCrudDeClientes.dto.ClientDTO;
 import com.highlander.DesafioCrudDeClientes.entities.Client;
 import com.highlander.DesafioCrudDeClientes.repositories.ClientRepository;
+import com.highlander.DesafioCrudDeClientes.services.exceptions.DatabaseException;
 import com.highlander.DesafioCrudDeClientes.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -39,17 +43,28 @@ public class ClientService {
     }
 
     @Transactional
-    public ClientDTO update(Long id , ClientDTO dto) {
-        Client entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
+    public ClientDTO update(Long id, ClientDTO dto) {
+        try {
+            Client entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
 
-        return new ClientDTO(entity);
+            return new ClientDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Referential integrity exception");
+        }
     }
 
     private void copyDtoToEntity(ClientDTO dto, Client entity) {
